switched email_id to bigint

+++ database/migrations/2016_06_29_160432_create_attribution_expiration_schedules_table.php
@@ -16,7 +16,7 @@ class CreateAttributionExpirationSchedulesTable extends Migration
     public function up()
     {
         Schema::create('attribution_expiration_schedules', function (Blueprint $table) {
-            $table->integer( 'email_id' )
+            $table->bigInteger( 'email_id' )
             $table->date( 'expiration_date' );
             $table->timestamps();
